Move AcvtionDescriptor to its own header
