+++ src/mapping_core/scenario.h
@@ -1,5 +1,6 @@
 #ifndef SCENARIO_H
 #define SCENARIO_H
+#include "action_descriptor.h"
 #include "basics.h"
 #include "escape_strings.h"
 #include "chk.h"
@@ -9,24 +10,6 @@
 #include <string>
 #include <map>
 #include <vector>
-
-// TODO: Header for these...
-
-enum class ActionDescriptor
-{
-    None = 0,
-    UpdateUnitSel,
-    ClearUnitSel,
-    CreateUnit,
-    PasteUnits,
-    BrushIsom
-};
-struct DescriptorIndex {
-    ActionDescriptor descriptorIndex = ActionDescriptor::None;
-    constexpr DescriptorIndex() noexcept = default;
-    constexpr DescriptorIndex(ActionDescriptor actionDescriptor) noexcept : descriptorIndex(actionDescriptor) {}
-    friend constexpr bool operator==(const DescriptorIndex & lhs, const DescriptorIndex & rhs) noexcept { return lhs.descriptorIndex == rhs.descriptorIndex; }
-};
 
 struct Scenario : RareEdit::Tracked<MapData, Scenario, DescriptorIndex>
 {
